Sharing is caring
The share button is working. And by working I mean that the mail
sharing is working. We will need to fix this if we are going to enable
messaging or any other share type. We may also need to restrict more
depending on what pops up when we check the real phonee.