White "blank" screen on init
Got it to be a white blank screen instead of the gray. In addition,
there were some issues when going to the more section and then back
that you had to select a different trial if you wanted to click on the
same one that you were viewing before. It still is highlighted, which
is another problem, but you can now reselect it to see it instead of
having select another first.